requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
drf-yasg==1.21.11
aiogram==3.22.0
requests==2.32.5
djangorestframework_simplejwt==5.5.1
